Fixed the test 'test_created_crs_in_georeferencer' in TestCRSCreator (#293)
The test 'test_created_crs_in_georeferencer' in TestCRSCreator was not working. It wasn't working because we were opening the geo reference before we loaded in the dataset, so the geo referencer did not have access to the dataset. So everytime we tried to click, we clicked on nothing. I also added to try_load_file a failure if the file doesn't exist.

Closes #292