Windows download and sign works (#257)
Windows download and code sign currently works. Mac codesign is being developed. The issue is that in the branch feat/build-on-runners2 we only get the folder with _internal and WISER_Bin and we don't get the .app folder which we need to download, sign, and build the distributable. Closes #296 and #295

* Windows download and sign works

* Added Makefile codesign command, and tested it
out to the extent that we can download the artifact.
However, we didn't originally upload the .app of
the artifact so we can't fully test the process of
downloading, signing, and packaging the app.

* Updated sign_windows so that it expects the downloaded artifact to
have everything in the dist folder (as opposed to
everything in the dist/WISER folder). sign_mac already
makes this assumption so no changes need there.

* Fixed default artifact names to download

* Got MakeFile recipe to download the artifact from github
but I ran into an issue with the zipping of the application
messing up symlinks, so i need to tar the application. I made
the github action upload tar it instead

* Added better way to find execuatable files
based off of Mach-O header

* Temporarily removing smoke tests for mac
so I can more easily test the tar zip and upload step

* Attempting to fix the tar command to put the .tgz in dist

* Added ability to use the .tar compressed
filed to codesign application

* Added back the other OS versions: windows and intel.

Added back smoke tests to mac

* Added more explanatory comments

* Aded print out to see what cv2_binaries aren't getting included

* got rid of saving to debug_output.txt

* Testing out cv2 libraries again

* Formatted ruff

* Attempt at fixing github build opencv error

* Removed debug_output.txt file

* Trying to see why the dynamic_lib path gets

added wrong to the path on the runner. It is getting
added as an absolute path to sys.path on the user's machine

* Using runtime hook to fix build machine cv2 path dependency

* ONly using windows to do the buiilds

* Removing collect_dynamic_libs for cv2 to ensure loading cv2

still works without it and with the runtime hook

* Adding back in collecting the cv2 libraries to be defensive

* Remade all of the operating systems build and run

on the github runners again.

* CLeaned up code comments and other unncessary lines