fix: Fixed excess slots added to cbox and sbox
Everytime _sync_band_choosers was called, a slot would be
added to each cbox and sbox for the band choosing. This
was not the functionality we wanted, we only wanted
these slots to be added once. Also everytime one dataset was
updated, the other one would be reset. This was very annoying
for the user