fix: Fixed bad slot logic for band chooser combo and spin boxes (#332)
## What does this change do?
Everytime _sync_band_choosers was called, a slot would be added to each
cbox and sbox for the band choosing. This was not the functionality we
wanted, we only wanted these slots to be added once. Also everytime one
dataset was updated, the other one would be reset. This was very
annoying for the user. This change fixes these issues. Closes #331 and
closes #333

## What type of PR is this? (check all applicable) 
- [ ] Feature
- [x] Bug Fix
- [ ] Documentation Update
- [ ] Style
- [ ] Code Refactor
- [ ] Performance Improvements
- [ ] Test
- [ ] Hot Fix
- [ ] Build
- [ ] CI
- [ ]
[Chore](https://stackoverflow.com/questions/26944762/when-to-use-chore-as-type-of-commit-message?utm_source=chatgpt.com)
- [ ] Revert

## Why is this change needed?
Look at the summary.

## How did you implement the change?
I fixed us constantly adding the same slots slots to the same combo
boxes and spi nboxes by only doing this add in the initialization. I
fixed the issue where the band chooser would be reset if a dataset for
the other band was changed by putting the sync_band_choosers logic into
two functions: one for x and one for y.

## Updated tests?
- [X] yes
- [ ] no, because they aren’t needed
- [ ] no, because I need help

## Added to documentation?  
- [ ] yes
- [x] no documentation needed 

## Checklist
- [x] There is an issue associated with this pull request
- [x] Code compiles/builds without errors
- [x] No new lint/style issues introduced
- [x] Branch is up-to-date with main/master
- [x] All CI checks passed

## Desktop Screenshots/Recordings  
<!-- If applicable, add screenshots or video links showing changes. -->

## [optional] Are there any post-merge tasks we need to perform?  
<!-- List any tasks required after merge. -->