feat: Creation of common UI library class (#327)
## What does this change do?
This change introduces the basics of the new UI library. So doing things
like making dataset, spectrum, roi, and band choosers. Works towards
#322 . It adds 4 new functions to Application state that are mean to
give plugin creators UI pieces for easily getting input. The functions
are: `choose_dataset_ui`, `choose_spectrum_ui`, `choose_roi_ui`,
`choose_band_ui`.

## What type of PR is this? (check all applicable) 
- [x] Feature
- [ ] Bug Fix
- [ ] Documentation Update
- [ ] Style
- [ ] Code Refactor
- [ ] Performance Improvements
- [ ] Test
- [ ] Hot Fix
- [ ] Build
- [ ] CI
- [ ]
[Chore](https://stackoverflow.com/questions/26944762/when-to-use-chore-as-type-of-commit-message?utm_source=chatgpt.com)
- [ ] Revert

## Why is this change needed?
This change is needed so it is easier for plugin makers to create the
GUI pieces to take in a dataset or spectrum or ROI or image band to do
whatever math or visualizations that they need to do for it.

## How did you implement the change?
<!-- High-level approach -->

## Added tests?
- [X] yes
- [ ] no, because they aren’t needed
- [ ] no, because I need help

## Added to documentation?  
- [x] yes
- [ ] no documentation needed 

## Checklist
- [x] There is an issue associated with this pull request (#322)
- [x] Code compiles/builds without errors
- [x] No new lint/style issues introduced
- [x] Branch is up-to-date with main/master
- [x] All CI checks passed

## Desktop Screenshots/Recordings  
<!-- If applicable, add screenshots or video links showing changes. -->

## [optional] Are there any post-merge tasks we need to perform?  
<!-- List any tasks required after merge. -->