feat: Added plugin ui that lets people enter things (#334)
## What does this change do?
This adds a UI create for plugin users that lets them create a GUI with
line edits and combo boxes. Works towards #322

## What type of PR is this? (check all applicable) 
- [x] Feature
- [ ] Bug Fix
- [ ] Documentation Update
- [ ] Style
- [ ] Code Refactor
- [ ] Performance Improvements
- [ ] Test
- [ ] Hot Fix
- [ ] Build
- [ ] CI
- [ ]
[Chore](https://stackoverflow.com/questions/26944762/when-to-use-chore-as-type-of-commit-message?utm_source=chatgpt.com)
- [ ] Revert

## Why is this change needed?
This change is needed so plugin developers can prompt users for more
than just objects in WISER.

## How did you implement the change?
I made it so the code loops through a list of passed in parameters and
constructs the dialog from that list.
Because the dialog has to be made from the function, we can't call exec_
on the dialog because the dialog would not be populated with widgets.

## Added tests?
- [ ] yes
- [x] no, because they aren’t needed
- [ ] no, because I need help

## Added to documentation?  
- [x] yes
- [ ] no documentation needed 

## Checklist
- [x] There is an issue associated with this pull request (#322)
- [x] Code compiles/builds without errors
- [x] No new lint/style issues introduced
- [x] Branch is up-to-date with main/master
- [x] All CI checks passed

## Desktop Screenshots/Recordings  
Here is an example of a UI piece I made:
<img width="206" height="82" alt="image"
src="https://github.com/user-attachments/assets/e6e2bf88-780b-4066-ac4c-81160e5621c5"
/>
And the plugin code I used to make it
```python
        form_inputs = [
            ("Enter Wavelength", "wvl", 2),
            ("Enter Divisor", "divisor", 1),
            ("Enter Dimensionality Reduction", "dim_reduction", 0, ["PCA", "SVD"]),
        ]
        return_dict: Dict[str, Any] = self._app_state.create_form(
            form_inputs,
            title="Analysis Params",
            description="In the above text enter the parameters needed to perform the algorithm.",
        )
```

## [optional] Are there any post-merge tasks we need to perform?  
<!-- List any tasks required after merge. -->