feat: Better interface with RasterDataSet and RasterData and more plugin UI library objects (#336)
## What does this change do?
This adds the `TableDisplayWidget`, `MatplotlibDisplayWidget` classes to
the UI library and a way for plugin users to make a
`SpectrumPlotGeneric` class through app_state. The classes
`TableDisplayWidget`, `MatplotlibDisplayWidget` are also exposed to
plugin users through app_state. This also adds tests for
`TableDisplayWidget`. It also adds documentation on using these 3 . I
had to add functions to RasterDataSet and RasterDataBand to expose
values that would be useful to the user (like data ignore) and setting
band_info. I also had to redo some imports to put them under type
checking to fix circular import issues.

Addresses #322

Closes #337

## What type of PR is this? (check all applicable) 
- [X] Feature
- [ ] Bug Fix
- [X] Documentation Update
- [ ] Style
- [ ] Code Refactor
- [ ] Performance Improvements
- [X] Test
- [ ] Hot Fix
- [ ] Build
- [ ] CI
- [ ]
[Chore](https://stackoverflow.com/questions/26944762/when-to-use-chore-as-type-of-commit-message?utm_source=chatgpt.com)
- [ ] Revert

## Why is this change needed?
This change is needed to wrap up the current work on making it easier
for users to create plugins. We needed this because we want the barrier
of entry to make plugins to be lower.

## How did you implement the change?
I made 2 new classes and 3 new functions in app_state to expose 3
classes.

## Added tests?
- [X] yes (not comprehensive)
- [ ] no, because they aren’t needed
- [ ] no, because I need help

## Added to documentation?  
- [X] yes
- [ ] no documentation needed 

## Checklist
- [X] There is an issue associated with this pull request
- [X] Code compiles/builds without errors
- [x] No new lint/style issues introduced
- [x] Branch is up-to-date with main/master
- [x] All CI checks passed

## Desktop Screenshots/Recordings  
<!-- If applicable, add screenshots or video links showing changes. -->

## [optional] Are there any post-merge tasks we need to perform?  
<!-- List any tasks required after merge. -->