feat: Adding in sam and sff on image (#346)
## What does this change do?
This PR adds the ability to do SAM and SFF on an image cube for
reference spectra. It also fixes a bug in continuum removal. It also
adds the ability to use collected and active spectrum in SAM and SFF. It
also redoes the SAM and SFF single algorithm data processing to match
the image cube algorithm. Closes #352 Closes #353

SAM data processing:
- Target:
  - Sliced target to bounds of wavelength range
- Mask target based on finite values of both target and reference and
the target and reference
bad bands (all & together) (all of these arrays have 1's for keep and
0's for ignore)
- Reference:
- Resample (with 1D interpolation) reference spectrum and wavelengths
and bad bands to target wavelengths
- Slice reference spectrum and wavelengths and bad bands to wavelength
range
- Mask target based on finite values of both target and reference and
the target and reference
bad bands (all & together) (all of these arrays have 1's for keep and
0's for ignore)
- Computing
  - Compute sine angle

SFF data processing:
- Target:
  - Slice target to bounds of wavelength range
  - Removed bad target bands from target
  - Compute continuum removal
- Reference:
  - Resample (1D interpolation)
  - Slice to bounds of wavelength range
  - Remove bad reference bands from reference
  - Compute continuum removal
- Remake array into the grid of the target when it was sliced, and make
the fill value for the reference bad
  band value one that won't affect the result (which is 1.0)
  - Remove the target's bad bands from reference
- Computing
  - Compute SFF

Continuum Removal Fix:
- The algorithm for continuum removal that a previous intern wrote was
wrong. It reversed the wavelength array but did not reverse the
spectrum. This by itself does not affect the continuum, as it introduced
a negative into the dot product equation and they introduced another
negative by using the wrong direction of the dot product. Since the math
was multiplicative, this allowed them to get the correct convex hull.
However, the issue occurred when they were interpolating based on the
convex hull. Since the wavelengths were not equally spaced, this cause
the interpolated values to be incorrect. Tests currently ensure this
doesn't have a regression.

## What type of PR is this? (check all applicable) 
- [X] Feature
- [x] Bug Fix
- [ ] Documentation Update
- [ ] Style
- [ ] Code Refactor
- [X] Performance Improvements
- [X] Test
- [ ] Hot Fix
- [ ] Build
- [ ] CI
- [ ]
[Chore](https://stackoverflow.com/questions/26944762/when-to-use-chore-as-type-of-commit-message?utm_source=chatgpt.com)
- [ ] Revert

## Why is this change needed?
We need SAM and SFF to be able to operate on images and compare spectra
from the images so it is easy for people to run the algorithms on their
datasets.

## How did you implement the change?
<!-- High-level approach -->

## Added tests?
- [x] yes
- [ ] no, because they aren’t needed
- [ ] no, because I need help

## Added to documentation?  
- [ ] yes
- [x] no documentation needed 

## Checklist
- [x] There is an issue associated with this pull request (#352)
- [x] Code compiles/builds without errors
- [x] No new lint/style issues introduced
- [x] Branch is up-to-date with main/master
- [x] All CI checks passed

## Desktop Screenshots/Recordings  
<!-- If applicable, add screenshots or video links showing changes. -->

## [optional] Are there any post-merge tasks we need to perform?  
<!-- List any tasks required after merge. -->