feat: Preventing sam & sff from bloating in memory. Fix nans with certain spectra (#354)
## What does this change do?
Prevents sam and sff from recomputing a 3D image cube on each iteration
of a for loop. It also fixes nans propagating through and messing up our
data. Closes #355 Closes #356

## What type of PR is this? (check all applicable) 
- [ ] Feature
- [X] Bug Fix
- [ ] Documentation Update
- [ ] Style
- [ ] Code Refactor
- [X] Performance Improvements
- [X] Test
- [ ] Hot Fix
- [ ] Build
- [ ] CI
- [ ]
[Chore](https://stackoverflow.com/questions/26944762/when-to-use-chore-as-type-of-commit-message?utm_source=chatgpt.com)
- [ ] Revert

## Why is this change needed?
This change is needed so that we keep a low memory overhead and have
fast computation when performing parallel runs of sam and sff for
multiple reference spectra.

## How did you implement the change?
I redid some of the math to accumulate values in order to keep the low
memory foot print.
For the nan's propagating, I explicitly set all non-finite values to 0.0
before doing the boolean cast.

## Added tests?
- [X] yes
- [ ] no, because they aren’t needed
- [ ] no, because I need help

## Added to documentation?  
- [ ] yes
- [X] no documentation needed 

## Checklist
- [X] There is an issue associated with this pull request (#355) (#356)
- [X] Code compiles/builds without errors
- [X] No new lint/style issues introduced
- [X] Branch is up-to-date with main/master
- [X] All CI checks passed

## Desktop Screenshots/Recordings  
<!-- If applicable, add screenshots or video links showing changes. -->

## [optional] Are there any post-merge tasks we need to perform?  
<!-- List any tasks required after merge. -->